split index to app&database
